refactor: move components to ui directory
