+++ src/app/(homepage)/page.tsx
@@ -1,4 +1,4 @@
-import { FamilyMemberCard } from '@local/components/FamilyMemberCard';
+import { FamilyMemberCard } from '@local/ui/FamilyMemberCard';
 import styles from '@local/styles/Home.module.scss';
 
 export default function HomepagePage() {
