+++ src/ui/Header/Header.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styles from './Header.module.scss';
-import { Navigation } from '@local/components/Navigation';
+import { Navigation } from '@local/ui/Navigation';
 
 export default function Header() {
   return (
